+++ libs/langchain/src/agents/nodes/ToolNode.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-instanceof/no-instanceof */
 import {
   BaseMessage,
   ToolMessage,
@@ -22,10 +23,13 @@ import {
 } from "@langchain/langgraph";
 
 import { RunnableCallable } from "../RunnableCallable.js";
-import { PreHookAnnotation } from "../annotation.js";
-import { mergeAbortSignals } from "../utils.js";
+import {
+  PreHookAnnotation,
+  AnyAnnotationRoot,
+  ToAnnotationRoot,
+} from "../annotation.js";
+import { mergeAbortSignals } from "./utils.js";
 import { ToolInvocationError } from "../errors.js";
-import type { AnyAnnotationRoot, ToAnnotationRoot } from "../types.js";
 
 export interface ToolNodeOptions {
   /**
