+++ libs/langchain/src/hub/tests/hub.int.test.ts
@@ -1,3 +1,4 @@
+/* eslint-disable no-process-env */
 import { expect, test } from "vitest";
 
 import { ChatPromptTemplate, PromptTemplate } from "@langchain/core/prompts";
